Move MonadLift into its own module.
